move common variables/functions into a utils file; read backend url from env file
