+++ frontend/src/env.example
@@ -1,0 +1,1 @@
+REACT_APP_BACKEND_URL="http://127.0.0.1:8000"
